refactor(web-client): Project restructuring
Defining structure of client and api
