+++ client/.env
@@ -1,0 +1,4 @@
+REACT_APP_API_ENDPOINT=http://localhost:3000
+REACT_APP_NAME=Telzir
+REACT_APP_ENV=development
+PORT=3001
