+++ client/.eslintrc.json
@@ -1,0 +1,23 @@
+{
+	"env": {
+		"browser": true,
+        "commonjs": true,
+		"es6": true
+	},
+	"extends": "react-app",
+	"parserOptions": {
+        "ecmaFeatures": {
+            "jsx": true
+        },
+        "ecmaVersion": 2018
+	},
+	"rules": {
+		"indent": ["warn", "tab"],
+		"quotes": ["error", "single"],
+		"no-unused-vars": "warn",
+		"no-undef": "warn",
+		"semi": ["error", "always"],
+		"no-console": ["error", { "allow": ["warn", "error", "log"] }]
+    },
+    "plugins": ["import"]
+}
