+++ client/.prettierrc
@@ -1,0 +1,8 @@
+{
+    "singleQuote": true,
+    "useTabs": true,
+    "printWidth": 105,
+    "tabWidth": 2,
+    "cssEnable": ["scss"],
+    "eslintIntegration": true
+}
