+++ client/README.md
@@ -1,0 +1,15 @@
+# Telzir WebClient
+
+## Setup
+
+Install dependencies
+
+```bash
+$ npm i
+```
+
+Run Development evironment
+
+```bash
+$ npm start
+```
