+++ client/package.json
@@ -3,9 +3,13 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "axios": "0.18.0",
+    "node-sass": "4.11.0",
     "react": "^16.7.0",
     "react-dom": "^16.7.0",
-    "react-scripts": "2.1.3"
+    "react-scripts": "2.1.3",
+    "styled-components": "4.1.3",
+    "styled-media-query": "2.1.1"
   },
   "scripts": {
     "start": "react-scripts start",
