+++ client/src/components/template/footer/package.json
@@ -1,0 +1,3 @@
+{
+	"main": "Footer.jsx"
+}
