+++ client/src/components/template/header/Header.jsx
@@ -1,0 +1,21 @@
+import React, { Component } from 'react';
+// import styled from 'styled-components';
+
+/**
+ * Header main component
+ *
+ * @class Header
+ * @extends {Component}
+ */
+class Header extends Component {
+	/**
+	 * Component render method
+	 *
+	 * @returns
+	 * @memberof Header
+	 */
+	render() {
+		return <div>Header</div>;
+	}
+}
+export default Header;
