+++ client/src/components/template/header/package.json
@@ -1,0 +1,3 @@
+{
+	"main": "Header.jsx"
+}
