+++ client/src/components/template/index.js
@@ -1,0 +1,2 @@
+export { default as Footer } from './footer/';
+export { default as Header } from './header/';
