+++ client/src/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
+import './main.scss';
 import App from './App';
 import * as serviceWorker from './serviceWorker';
