+++ client/src/main.scss
@@ -1,0 +1,11 @@
+@charset "UTF-8";
+
+@import "./scss/components/color-variables";
+@import "./scss/components/variables";
+@import "./scss/main";
+@import "./scss/components/color-classes";
+@import "./scss/components/reset";
+@import "./scss/components/typography";
+@import "./scss/components/grid";
+@import "./scss/components/buttons";
+@import "./scss/components/cards";
