+++ client/src/scss/_main.scss
@@ -1,0 +1,233 @@
+html,
+body {
+	font-family: Roboto light, sans-serif;
+	margin: 0;
+	padding: 0;
+	display: flow-root;
+	font-size: 14px;
+	background-color: color('grey', 'lighten-5');
+}
+.left-align,
+.pull-left {
+	left: 0;
+	text-align: left;
+	float: left;
+	justify-content: left;
+}
+
+.right-align,
+.pull-right {
+	right: 0;
+	float: right;
+	text-align: right;
+	justify-content: right;
+}
+
+.center,
+.center-align {
+	text-align: center !important;
+	align-items: center !important;
+	justify-content: center !important;
+}
+
+.left {
+	float: left !important;
+}
+
+.right {
+	float: right !important;
+} // No Text Select
+.no-select {
+	user-select: none;
+}
+
+.circle {
+	border-radius: 50%;
+}
+
+.center-block {
+	display: block;
+	margin-left: auto;
+	margin-right: auto;
+}
+
+.truncate {
+	display: block;
+	white-space: nowrap;
+	overflow: hidden;
+	text-overflow: ellipsis;
+}
+
+.no-padding {
+	padding: 0 !important;
+}
+
+.field-error-alert {
+	color: red !important;
+	font-style: italic !important;
+}
+
+// Z-levels
+.z-depth-0 {
+	box-shadow: none !important;
+}
+
+/* 2dp elevation modified*/
+.z-depth-1 {
+	box-shadow: 0 2px 2px 0 rgba(0, 0, 0, 0.14), 0 3px 1px -2px rgba(0, 0, 0, 0.12),
+		0 1px 5px 0 rgba(0, 0, 0, 0.2);
+}
+.z-depth-1-half {
+	box-shadow: 0 3px 3px 0 rgba(0, 0, 0, 0.14), 0 1px 7px 0 rgba(0, 0, 0, 0.12),
+		0 3px 1px -1px rgba(0, 0, 0, 0.2);
+}
+
+/* 6dp elevation modified*/
+.z-depth-2 {
+	box-shadow: 0 4px 5px 0 rgba(0, 0, 0, 0.14), 0 1px 10px 0 rgba(0, 0, 0, 0.12),
+		0 2px 4px -1px rgba(0, 0, 0, 0.3);
+}
+
+/* 12dp elevation modified*/
+.z-depth-3 {
+	box-shadow: 0 8px 17px 2px rgba(0, 0, 0, 0.14), 0 3px 14px 2px rgba(0, 0, 0, 0.12),
+		0 5px 5px -3px rgba(0, 0, 0, 0.2);
+}
+
+/* 16dp elevation */
+.z-depth-4 {
+	box-shadow: 0 16px 24px 2px rgba(0, 0, 0, 0.14), 0 6px 30px 5px rgba(0, 0, 0, 0.12),
+		0 8px 10px -7px rgba(0, 0, 0, 0.2);
+}
+
+/* 24dp elevation */
+.z-depth-5 {
+	box-shadow: 0 24px 38px 3px rgba(0, 0, 0, 0.14), 0 9px 46px 8px rgba(0, 0, 0, 0.12),
+		0 11px 15px -7px rgba(0, 0, 0, 0.2);
+}
+
+// Tables
+table,
+th,
+td {
+	border: none;
+}
+
+table {
+	width: 100%;
+	display: table;
+	border-collapse: collapse;
+	border-spacing: 0;
+
+	&.striped {
+		tr {
+			border-bottom: none;
+		}
+
+		> tbody {
+			> tr:nth-child(odd) {
+				background-color: $table-striped-color;
+			}
+
+			> tr > td {
+				border-radius: 0;
+			}
+		}
+	}
+
+	&.highlight > tbody > tr {
+		transition: background-color 0.25s ease;
+		&:hover {
+			background-color: $table-striped-color;
+		}
+	}
+
+	&.centered {
+		thead tr th,
+		tbody tr td {
+			text-align: center;
+		}
+	}
+}
+
+tr {
+	border-bottom: 1px solid $table-border-color;
+}
+
+td,
+th {
+	padding: 15px 5px;
+	display: table-cell;
+	text-align: left;
+	vertical-align: middle;
+	border-radius: 2px;
+}
+
+// Responsive Table
+@media #{$medium-and-down} {
+	table.responsive-table {
+		width: 100%;
+		border-collapse: collapse;
+		border-spacing: 0;
+		display: block;
+		position: relative;
+
+		td:empty:before {
+			content: '\00a0';
+		}
+
+		th,
+		td {
+			margin: 0;
+			vertical-align: top;
+		}
+
+		th {
+			text-align: left;
+		}
+		thead {
+			display: block;
+			float: left;
+
+			tr {
+				display: block;
+				padding: 0 10px 0 0;
+
+				th::before {
+					content: '\00a0';
+				}
+			}
+		}
+		tbody {
+			display: block;
+			width: auto;
+			position: relative;
+			overflow-x: auto;
+			white-space: nowrap;
+
+			tr {
+				display: inline-block;
+				vertical-align: top;
+			}
+		}
+		th {
+			display: block;
+			text-align: right;
+		}
+		td {
+			display: block;
+			min-height: 1.25em;
+			text-align: left;
+		}
+		tr {
+			border-bottom: none;
+			padding: 0 10px;
+		}
+
+		/* sort out borders */
+		thead {
+			border: 0;
+			border-right: 1px solid $table-border-color;
+		}
+	}
+}
